Add version to CLI
