+++ .eslintrc.js
@@ -2,5 +2,8 @@ module.exports = {
   "extends": "airbnb-base",
   "rules": {
     "no-underscore-dangle": 0
+  },
+  "globals": {
+    VERSION: true
   }
 };
