+++ README.md
@@ -10,12 +10,11 @@ It runs on top of NodeJS and it's meant to unit-test JavaScript
 applications.
 
 ## Why you should use it?
-1. It has everything you need in most of the cases in only ~6Kb;
+1. It has everything you need in most of the cases in only ~8Kb;
 2. It has an almost flat learning curve if you're already familiar with
 other testing frameworks like Mocha, Jasmine, Jest;
 3. It has an almost flat learning curve even though you're not familiar
-with other testing technologies but you're familiar with ES6+
-(opinionated);
+with other testing technologies but you're familiar with ES6+;
 4. It's deadly simple. You can read the source and understand how it
 works in less than one hour;
 
@@ -43,23 +42,10 @@ Then you can use the command line interface this way:
 node_modules/.bin/titef test/awesome.specs.js
 ```
 
-Alternatively you can use it via `npm` scripts:
-
-```json
-// package.json
-{
-  ...
-  "scripts": {
-    "titef": "titef"
-  },
-  ...
-}
-```
-
-And then
+or [better](https://github.com/zkat/npx)
 
 ```bash
-npm run titef -- test/awesome.specs.js
+npx titef test/awesome.specs.js
 ```
 
 ## Getting started
