+++ lib/cli/index.js
@@ -1,14 +1,18 @@
-const { resolve } = require('path');
-const { existsSync } = require('fs');
-const printHelp = require('./_print-help');
+const {resolve} = require('path');
+const {existsSync} = require('fs');
+const {printHelp, printVersion} = require('./print');
 
 // Removes `node` and `__filename` from arguments list
 const nonTrivialArguments = process.argv.slice(2);
 
 const showHelp = nonTrivialArguments.find(item => item === '-h' || item === '--help');
+const showVersion = nonTrivialArguments.find(item => item === '-v' || item === '--version');
 
 if (showHelp) {
   printHelp();
+  process.exit(0);
+} else if (showVersion) {
+  printVersion();
   process.exit(0);
 }
